LayoutTests/fast/css/position-absolute-float.html

<!DOCTYPE html>
<html>
<head>
<style>
#static { position: static; float: right; }
#absolute { position: absolute; float: right; }
#fixed { position: fixed; float: right; }
</style>
<script src="../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<div id="static"></div>
<div id="absolute"></div>
<div id="fixed"></div>
</body>
<script>
description("Test for Bug: 105836 - computed value of float on positioned elements.");
shouldBe('getComputedStyle(document.getElementById("static")).float', '"right"');
shouldBe('getComputedStyle(document.getElementById("absolute")).float', '"none"');
shouldBe('getComputedStyle(document.getElementById("fixed")).float', '"none"');
</script>
<script src="../js/resources/js-test-post.js"></script>
</html>
