third_party/blink/web_tests/fast/css/position-absolute-float-expected.txt

Test for Bug: 105836 - computed value of float on positioned elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("static")).float is "right"
PASS getComputedStyle(document.getElementById("absolute")).float is "none"
PASS getComputedStyle(document.getElementById("fixed")).float is "none"
PASS successfullyParsed is true

TEST COMPLETE
